feat(scene3d): port meshes / glTF PBR to QRhi
The shading core is ported whole rather than reduced: base colour in both
modes (material-driven vs per-draw override), all five metallic-roughness
maps, normal mapping, the fixed world key light plus the camera-locked
fill, and the analytic image-based ambient — diffuse irradiance plus the
split-sum specular reflection of a procedural sky/ground environment,
which is what makes metals read as metal instead of near-black.

Two structural departures, both forced by QRhi rather than chosen:

- Every material binds all five samplers, because a pipeline is compiled
  against its binding layout and that layout cannot vary per material.
  Absent maps bind a 1x1 NEUTRAL texel, which also deletes four of GL's
  five u_has_*_tex flags: white is the identity for the multiplicative
  slots. The normal map keeps a flag, since its neutral value is a no-op
  only when the tangent basis is well-formed and a mesh without UVs has a
  degenerate one. The flag follows what was ACTUALLY bound, not what the
  material asked for, so a map that fails to decode degrades instead of
  corrupting shading.
- Per-draw uniforms ride a dynamic-offset UBO, as QRhi has no loose
  uniforms. Growing that buffer invalidates every binding set referencing
  it, so capacity is settled from an upper bound BEFORE resolving a
  frame's draws rather than after counting them.

Placeholder texture creation and upload are deliberately split: a
QRhiResourceUpdateBatch obtained inside initialize() is never submitted,
so uploading content there silently loses it.

While here, extracts the box/cylinder/sphere builders out of the GL pass's
anonymous namespace into mesh_primitives.{h,cpp} so both backends draw
identical shapes, and adds camera_pos_world to RhiFrameContext (view-
dependent shading needs it and it is not recoverable from view_proj
without inverting it).

Also clears 10 pre-existing -Wsign-conversion warnings across the four
earlier QRhi passes; the whole port is now warning-clean. Verified
behaviour-neutral: the demo render is pixel-identical across that change.

Shadow receive and the R8 is-mesh mask for EDL are not ported, because the
passes that produce those inputs do not exist on QRhi yet.

+++ pj_scene3D/widgets/src/rhi/rhi_grid_pass.cpp
@@ -115,7 +115,7 @@ void RhiGridPass::prepare(QRhiResourceUpdateBatch& updates, const RhiFrameContex
     // Immutable: the tessellation only changes when the user edits extent or
     // divisions, which re-creates the buffer rather than streaming it per frame.
     delete vbo_;
-    vbo_ = rhi_->newBuffer(QRhiBuffer::Immutable, QRhiBuffer::VertexBuffer, bytes);
+    vbo_ = rhi_->newBuffer(QRhiBuffer::Immutable, QRhiBuffer::VertexBuffer, static_cast<quint32>(bytes));
     if (vbo_ == nullptr || !vbo_->create()) {
       vertex_count_ = 0;
       return;
